fix: refactor int to double for invitation expiration

+++ app/src/main/java/com/android/joinme/model/invitation/InvitationsRepository.kt
@@ -17,7 +17,7 @@ interface InvitationsRepository {
       type: InvitationType,
       targetId: String,
       createdBy: String,
-      expiresInDays: Int? = null
+      expiresInDays: Double? = null
   ): Result<String>
 
   /**
